Migrate all Resources tests to new framework
