+++ azure-mgmt-resource/tests/recordings/test_mgmt_resource.test_providers.yaml
