+++ azure-mgmt-resource/tests/recordings/test_mgmt_resource_locks.test_locks.yaml
