+++ azure-mgmt-resource/tests/recordings/test_mgmt_resource_subscriptions.test_subscriptions.yaml
